2.22.2124
* Early Access UE5 support
** Mod Browser UI is still WIP, but core functionality through UModioSubsystem should be fully available
* Zip64 Support
* QueryCurrentModUpdate improvements on non-Windows platforms - FModioModProgressInfo objects should be more accurate
* NativeSDK Updated
* Bugfixes and improvements to the Mod Browser UI

+++ Source/ModioUI/Public/UI/BaseWidgets/Slate/SModioTileView.h
@@ -260,7 +260,7 @@ public:
 				// Only scroll the item into view if it's not already in the visible range
 				const int32 NumItemsPerLine = this->GetNumItemsPerLine();
 				const double ScrollLineOffset = this->GetTargetScrollOffset() / NumItemsPerLine;
-				const int32 LineOfItem = FMath::FloorToInt(IndexOfItem / NumItemsPerLine);
+				const int32 LineOfItem = FMath::FloorToInt((float)IndexOfItem / (float)NumItemsPerLine);
 				const int32 NumFullLinesInView =
 					FMath::FloorToInt(ScrollLineOffset + NumLinesInView) - FMath::CeilToInt(ScrollLineOffset);
